# Various AArch64 speed hacks