- Use inline assembly to force `UMADDL` in the scalarRound code
   - Prevents the slow 64-bit `MADD` which has +4c latency and a 2c stall on little cores.
   - Also prevents a mask
   - Improves scalar speed even on big cores, should also improve mixed NEON speed on those little cores, although I haven't tested it.
 - Use inline assembly for `vmlal_u32`
   - Fixes NEON performance being garbage on GCC < 11 because it can't fold `vget_low/high` into `vmlal` (arm_neon.h uses inline assembly)
   - Documents the presumed reason why umlal must come before add.
   - Doubles as Clang's asm guard. (The old asm guard is still used on the single vector path)
   - Much cleaner in the code itself
 - Fix header detection
   - Include arm_neon.h when `__ARM_FEATURE_SVE` is defined
   - Don't include arm_neon.h if only `__aarch64__` is defined (e.g. -march=armv8-a+nofp)
   - Don't include arm_neon.h on < ARMv7 MSVC (which only would happen on old WinCE toolchains)
 - Fix the generateSecret `MOVK` hack
   - Add it for GCC, it does that too now
   - Remove the assert because the change to use `unreachable()` undoes it
 - Make asm guards non-volatile to allow reordering
 - Fix GCC strict aliasing issues with both NEON and VSX (same bug). `__sync_synchronize()` is no longer necessary.