@easyaspi314 2023-03-03
Cortex-A53 results are in.

`XXH3_64b` 100KB benchmark

| Compiler | old XXH_SCALAR | old XXH_NEON | new XXH_SCALAR | new XXH_NEON |
|:--|:--|:--|:--|:--|
| GCC 12 | 1903 MB/s | 2263 MB/s | 2267 MB/s | 2450 MB/s |
| GCC 10 | 1825 MB/s | 1877 MB/s | 2289 MB/s | 2065 MB/s |
| Clang 15 | 1783 MB/s | 1946 MB/s | 2055 MB/s | 2035 MB/s |
| Clang 11 | 1717 MB/s | 1819 MB/s | 1917 MB/s | 1964 MB/s |