@easyaspi314 2023-03-03
The XXH32/XXH64 ordering thing interests me, as the optimizations vary wildly on the processor.

Just take XXH32.

| Kind | Speed (A53) | Speed (X1) |
| :-- | :-- | :-- | 
| NEON | 1.8 GB/s | 3.5 GB/s |
| Scalar, guard between rounds | 660 MB/s | 10 GB/s |
| Scalar, all guarded at end | 1.6 GB/s | 7.7 GB/s |