@Cyan4973 2023-03-03
> I wonder if the optimal play is to go for the happy medium. 🤔

It seems so