@easyaspi314 2023-03-05
Just realized that it could be possible for the compiler to generate the Cortex-A53 errata with the umaddl. My tablet supposedly has this errata and hasn't seen an issue, but the fix is to prevent it from putting a load/store before a multiply-accumulate. (Usually a `nop`). However, I have an idea that ensures this with the least penalty, and that is using that stall cycle to do the shift, which if I use the variable form, is always 1 cycle even on the A53 line. Still not great having the dependency there but I'd rather that than the multiply result randomly being corrupted.

----

Additionally, one frustrating thing about XXH32 is that the best codegen (outside of `-fno-tree;slp-vectorize`) is from putting all 4 guards at the bottom of the loop, which is no problem on AArch64 with its 31 GPRs, but x86 is probably going to have issues because of how few registers it has. Also guarding one variable isn't enough — NEON has half vectors and Clang will happily vectorize using 64-bit vectors which is even worse.