@easyaspi314 2023-03-05
Scratch that, I found out a better solution for XXH32. The asm guards don't need to be marked volatile. This was causing a reorder barrier, when we just needed to tell the compiler "put this in a register and mark it modified sometime between the two times it is accessed"