@easyaspi314 2023-03-05
That should do it. I removed the volatile from the asm guards and brought back the clang guards.

This does cause a slight slowdown on XXH32 for the X1, but a great benefit on the A53.

**Clang's XXH3 is now on par with GCC,** as it can interleave now.