@easyaspi314 2023-03-05
Also I'd like to see how the 6:2 works on Apple Silicon. With the reordering fixed it could actually be faster than 8 NEON.