@easyaspi314 2023-03-05
Update. I found my old SSH key, and the M1 still prefers 8 NEON lanes.

Since I needed to boot my old phone (Snap 730/Cortex-A76) anyways, I did a quick bench on it. GCC 12 had a slight performance regression, but it is only by like 4% (12.3 -> 11.8). This is despite the original code having the slower `madd`. 

Might be some ordering shenanigans that can be eventually worked out. Clang 15 did increase performance slightly (10.3 -> 10.8). Still not optimal though. This might be worth looking into in the future. 

Neither of these performance changes are particularly bad though.