@easyaspi314 2023-03-05
Killed two birds with one stone. The issue was in fact strict aliasing, and it was the exact same bug that was breaking s390x.

The accumulator vectors are now pointers to vectors with the `__attribute__((may_alias))` which excludes them from GCC's strict aliasing optimizations.