@easyaspi314 2023-03-10
> I made some measurements of this version on some ARM devices around, and here is what I found : 
> 
> | `arm64` cpu | vector | `dev` | `pr809` | diff |
> | --- | --- | --- | --- | --- |
> | `M1` | `NEON` | 35.5 GB/s | 35.5 GB/s | none |
> | `M1` | Scalar` | 20.0 GB/s | 22.2 GB/s | +11% |
> | `Snap8g1` | `NEON` | 25.5 GB/s | 26.6 GB/s | +4% |
> | `Snap8g1` | Scalar | 13.6 GB/s | 15.8 GB/s | +16% |
> 
> So this PR seems to provide more benefits to the Scalar mode.
> Is that the expectation ?

Yes. The benefits to scalar also applies to hybrid NEON, primarily on the in-order cores since they can't dispatch the scalar instructions for free.

It also affects GCC 10 and below. Since the implementation of `vmlal_u32` used inline assembly, GCC couldn't detect the pattern to merge the intrinsics. This caused it to output this garbage:
```unix-assembly
// (Literal vget_low_u32)
dup     d4, v1.d[0]
dup     d5, v2.d[0]
umlal   v0.2d, v4.2s, v5.2s
// (Literal vget_high_u32)
dup     d4, v1.d[1]
dup     d5, v2.d[1]
umlal   v3.2d, v4.2s, v5.2s
```
instead of the intended result of this:
```unix-assembly
umlal   v0.2d, v1.2s, v2.2s
umlal2  v3.2d, v1.4s, v2.4s
```