@easyaspi314 2023-03-11
I think I know the exact reason why `add` must come after `umlal`.

`umlal` isn't fast, and the compiler is aware of that in its cost model. Compared to most NEON instructions which execute in 2 cycles, `umlal` is a 5 cycle instruction and is additionally restricted to half of the NEON pipelines.

Normally, the most logical thing would be to put the slowest instructions last, so the compiler puts the `umlal`s at the very end of the loop.

However, on the pipeline level this is the issue. In the double width NEON path, the two `umlal(2)` instructions both share two dependencies, from the `uzp1`/`uzp2`, the only thing that isn't dependent being the `add`. Since there is nothing else to process, there is nothing to do on the other NEON pipelines. All it can do is wait, and this ends up in a scalarization:

|   F0   |   F1   |
|:--|:--|
|  add   |  uzp2   |
| umlal  | add |
| (latency) | (idle) |
| umlal2 | (idle) |
| (latency) | (idle) |

When the `umlal` is forced to come first, the compiler reorders it so that it comes *early* in the loop, alongside instructions that can execute on the other pipelines. This means that the pipelines are almost always processing data, and we are at maximum efficiency.